Finished Videos 6-10
In these videos, i added platforms for the character to move on, I made
it so the space bar make the Tofu jump, I made the enemy move around at
constant pace. In videos 9 and 10 I had trouble making a right bound
and making it so the tofu destroys the enemy.