Health variable
I wrote code to make it so when the tofu hits the pointy side of the
spikes it loses 1 of it 5 health hitpoints. Then when its all out of
healt hitpoints, then the tofu will be moved back to it starting
position. I had already starting writing this code so I just needed to
fix one minor error.